@@ -18,6 +18,7 @@
     "@types/node": "^12.7.2",
     "babel-loader": "^8.0.6",
     "babel-plugin-import": "^1.12.0",
+    "babel-polyfill": "^6.26.0",
     "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^3.2.0",
     "file-loader": "^4.2.0",
@@ -37,8 +38,10 @@
   },
   "dependencies": {
     "antd": "^3.21.4",
+    "axios": "^0.19.2",
     "customize-cra": "^0.5.0",
-    "react": "^16.9.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.13.1",
     "react-app-rewired": "^2.1.3",
     "react-dom": "^16.9.0",
     "react-router-dom": "^5.0.1"
